+++ website/docusaurus.config.js
@@ -8,7 +8,7 @@ module.exports = {
   plugins: ["docusaurus-plugin-sass"],
   i18n: {
     defaultLocale: "en",
-    locales: ["en", "fr", "ko"],
+    locales: ["en", "fr", "ko", "ja", "es"],
   },
   themeConfig: {
     googleAnalytics: {
@@ -32,7 +32,7 @@ module.exports = {
     navbar: {
       title: "Riverpod",
       logo: {
-        alt: "Riverpod logo",
+        alt: "Riverpod",
         src: "img/logo.png",
       },
       items: [
@@ -54,9 +54,8 @@ module.exports = {
       ],
     },
     footer: {
-
       logo: {
-        alt: "Riverpod logo",
+        alt: "Riverpod",
         src: "img/full_logo.svg",
       },
       links: [