+++ website/sidebars.js
@@ -33,10 +33,7 @@ module.exports = {
       type: "category",
       label: "Migration",
       collapsed: false,
-      items: [
-        "migration/0.13.0_to_0.14.0",
-        "migration/0.14.0_to_1.0.0"
-      ],
+      items: ["migration/0.13.0_to_0.14.0", "migration/0.14.0_to_1.0.0"],
     },
     {
       type: "category",
@@ -45,20 +42,17 @@ module.exports = {
         {
           type: "link",
           label: "Counter",
-          href:
-            "https://github.com/rrousselGit/river_pod/tree/master/examples/counter",
+          href: "https://github.com/rrousselGit/river_pod/tree/master/examples/counter",
         },
         {
           type: "link",
           label: "Todo list",
-          href:
-            "https://github.com/rrousselGit/river_pod/tree/master/examples/todos",
+          href: "https://github.com/rrousselGit/river_pod/tree/master/examples/todos",
         },
         {
           type: "link",
           label: "Marvel API",
-          href:
-            "https://github.com/rrousselGit/river_pod/tree/master/examples/marvel",
+          href: "https://github.com/rrousselGit/river_pod/tree/master/examples/marvel",
         },
       ],
     },
@@ -84,14 +78,12 @@ module.exports = {
         {
           type: "link",
           label: "Firebase Starter",
-          href:
-            "https://github.com/lohanidamodar/flutter_firebase_starter/tree/feature/riverpod",
+          href: "https://github.com/lohanidamodar/flutter_firebase_starter/tree/feature/riverpod",
         },
         {
           type: "link",
           label: "Time Tracking App (with Firebase)",
-          href:
-            "https://github.com/bizz84/starter_architecture_flutter_firebase",
+          href: "https://github.com/bizz84/starter_architecture_flutter_firebase",
         },
         {
           type: "link",
@@ -106,8 +98,7 @@ module.exports = {
         {
           type: "link",
           label: "Resocoder's Weather Bloc to Weather Riverpod",
-          href:
-            "https://github.com/campanagerald/flutter-bloc-library-v1-tutorial",
+          href: "https://github.com/campanagerald/flutter-bloc-library-v1-tutorial",
         },
         {
           type: "link",
@@ -116,13 +107,19 @@ module.exports = {
         },
         {
           type: "link",
-          label: "Firebase Authentication with Riverpod Following Flutter DDD Architecture Pattern",
+          label:
+            "Firebase Authentication with Riverpod Following Flutter DDD Architecture Pattern",
           href: "https://github.com/pythonhubpy/firebase_authentication_flutter_DDD",
         },
-         {
+        {
           type: "link",
           label: "Todo App with Backup and Restore feature",
           href: "https://github.com/TheAlphaApp/flutter_riverpod_todo_app",
+        },
+        {
+          type: "link",
+          label: "Integrating Hive database with Riverpod (simple example)",
+          href: "https://github.com/GitGud31/theme_riverpod_hive",
         },
       ],
     },
@@ -132,22 +129,30 @@ module.exports = {
       collapsed: false,
       items: [
         {
+          type: "category",
+          label: "All Providers",
+          collapsed: true,
+          items: [
+            "providers/provider",
+            "providers/state_notifier_provider",
+            "providers/future_provider",
+            "providers/state_provider",
+          ],
+        },
+        {
           type: "link",
           label: "riverpod",
-          href:
-            "https://pub.dev/documentation/riverpod/latest/riverpod/riverpod-library.html",
+          href: "https://pub.dev/documentation/riverpod/latest/riverpod/riverpod-library.html",
         },
         {
           type: "link",
           label: "flutter_riverpod",
-          href:
-            "https://pub.dev/documentation//flutter_riverpod/latest/flutter_riverpod/flutter_riverpod-library.html",
+          href: "https://pub.dev/documentation//flutter_riverpod/latest/flutter_riverpod/flutter_riverpod-library.html",
         },
         {
           type: "link",
           label: "hooks_riverpod",
-          href:
-            "https://pub.dev/documentation/hooks_riverpod/latest/hooks_riverpod/hooks_riverpod-library.html",
+          href: "https://pub.dev/documentation/hooks_riverpod/latest/hooks_riverpod/hooks_riverpod-library.html",
         },
       ],
     },